refactor(analysis): mechanical split of analysis/analyzer.py into sibling modules
MAJ-04 module 2/6: analysis/analyzer.py 2031 -> 263 lines (facade).

- analysis_extraction.py: regex-based structural extraction (leaf)
- analysis_statistics.py: scipy-gated statistics/distributions/correlations
- analysis_complexity.py: complexity/maintainability/technical-debt metrics
- simulation_visualizations.py: simulation-result and cross-framework plots
- framework_comparison.py: cross-framework comparison and reporting
- analyzer.py: facade re-exporting all 30 moved names plus 4 retained
  entrypoints (perform_statistical_analysis, run_performance_benchmarks,
  perform_model_comparisons, generate_analysis_summary)

Pattern: 3.3.0 execute/processor.py mechanical split (12a565b2f).
2001 lines verified byte-identical (moved + kept) against the pre-split
file; only import headers differ. Availability blocks moved with their
single consumers and are re-exported (SCIPY_AVAILABLE, stats,
SEABORN_AVAILABLE, sns). Package __init__ and all consumers unchanged.

Verified: ruff 0 violations; mypy clean; facade contract (34 names incl.
test-imported _extract_simulation_metrics) and import probes pass;
targeted pytest tests/analysis + pipeline analyze + llm: 482 passed.
